ref #184 - toevoegen van safety checks voor bookmark manipulaties

+++ app/Http/Controllers/Web/Articles/BookmarkController.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Http\Controllers\Web\Articles;
 
 use App\Models\Article;
+use Illuminate\Contracts\Support\Renderable;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Spatie\RouteAttributes\Attributes\Get;
@@ -16,7 +17,9 @@ final readonly class BookmarkController
     #[Get(uri: 'bookmark/{article}', name: 'bookmark:create')]
     public function store(Request $request, Article $article): RedirectResponse
     {
-        $request->user()->bookmarks()->attach($article);
+        if ($request->user()->bookmarks->doesntContain($article)) {
+            $request->user()->bookmarks()->attach($article);
+        }
 
         return redirect()->action(DictionaryArticleController::class, $article);
     }
@@ -24,7 +27,9 @@ final readonly class BookmarkController
     #[Get(uri: 'unbookmark/{article}', name: 'bookmark:remove')]
     public function delete(Request $request, Article $article): RedirectResponse
     {
-        $request->user()->bookmarks()->detach($article);
+        if ($request->user()->bookmarks->contains($article)) {
+            $request->user()->bookmarks()->detach($article);
+        }
 
         return redirect()->action(DictionaryArticleController::class, $article);
     }
